+++ src/main/java/org/boot/growup/source/board/service/impl/ReplyServiceImpl.java
@@ -1,19 +1,19 @@
-package org.boot.growup.source.board.service;
+package org.boot.growup.source.board.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.boot.growup.source.admin.persist.entity.Admin;
 import org.boot.growup.source.board.dto.request.PostReplyRequestDTO;
-import org.boot.growup.source.board.persist.repository.ReplyRepository;
 import org.boot.growup.source.board.persist.entity.Inquiry;
 import org.boot.growup.source.board.persist.entity.Reply;
+import org.boot.growup.source.board.persist.repository.ReplyRepository;
+import org.boot.growup.source.board.service.ReplyService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
 public class ReplyServiceImpl implements ReplyService {
     private final ReplyRepository replyRepository;
